day_nine/backend/requirements.txt
flask
flask-sqlalchemy
flask-pydantic
flask-cors
pydantic[email]
sqlalchemy
psycopg2-binary
alembic
pyjwt
passlib[bcrypt]
python-dotenv
pytest
pytest-flask
